Change the script to allow being ran from anywhere.
Changed the script to be able to run from locations other than the .git
repository by using the script_path variable. This is obtained using the path
of __file__.

Also changed the argparser to store_true for the arguments that
should be True when they are set.

Also changed it to use "{}".format(arg) for the strings for consistency
since I used those for creating the new shell commands.

Added a missing requirement of netaddr (eyewitness) and changed
smtplib => secure-smtplib.