Add installation functionality and recall the amass() or subfinder() methods.
Now the application will ask you if you would like to install the tools using
the normal method of `os.system('go get -u -v <repo>')` and once completed
will recall the corresponding method we are in, to run the tool. Added logic to
also only retry once if we already attempted to install prior.